requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
tqdm>=4.65.0
opencv-python>=4.8.0
scipy>=1.10.0
scikit-learn>=1.2.0
onnx>=1.14.0
onnxruntime>=1.15.0
